add userSlice
