+++ src/redux/store.ts
@@ -1,13 +1,15 @@
-import {configureStore, Action} from "@reduxjs/toolkit";
+import {configureStore} from "@reduxjs/toolkit";
 import {featureSlice} from "./reducers/featureSlice";
 import {carSlice} from "./reducers/carSlice";
 import {requestSlice} from "./reducers/requestSlice";
+import {userSlice} from "./reducers/userSlice";
 
 const store = configureStore({
     reducer: {
         allFeatures: featureSlice.reducer,
         cars: carSlice.reducer,
-        requests: requestSlice.reducer
+        requests: requestSlice.reducer,
+        user: userSlice.reducer
     }
 });
